requirements.txt
pdfminer3k
pdf2image>=1.11.0
numpy
opencv-python
logzero